feat: reordering cols and lanes stored in db
Must use queryRawUnsafe because prisma cannot take whole sql statements
in a variable. It must be like
await prisma.queryRaw`select * from KanbanColumn`
where things like select, * and KanbanColumn also cannot be in a
variable.
This is fine as is, but for this use case where we have an unknown
number of headers to update, have to build the sql query