build: seperate field type data into 2 tables ->
for checkbox and drop down, IRL, I would incrase this number to 10 and
have check for any data which had more than 4 options, I would have a
datatype with just option and all the data, then also make sure I
remember to include this in the possible datatype in the code after.
In the meantime, after this branch has been deployed to other servers
then have a data engineer manually update those with more than 10
options.